access useEditorModalState() directly cuz no provider exists! js_only
cleanup and reorder imports
